meta: Revert "main" to "./lib/index", no dist
We don't have the `dist/` dir on `master` branch.

If package.json "main" field points to this non-existent folder,
then the package is effectively broken.

This commit restores "main" to point to "./lib/index", as this seems
to be working now in my tests.

We can pre-compile the `dist/` folder again when tagging the next
release, if we really want to, but I'm pretty sure Pulsar compiles
the code on first run, so the stakes for doing so seem pretty small
to me. (Impact should be limited to improvements in first startup
speed, if I understand correctly.)

---

Note: according to Mauricio, the command to build the `dist/` dir is:
`ELECTRON_VERSION=12 babel lib -d dist --compact auto --source-type module`

But this is not trivially obvious to me how to automate at the moment,
so I won't try to do so for now. This just gets the repo working on
`master` branch again, without bothering with pre-compilation stuff.