lib: Use nullish coalescing '??', not or '||'
Per review suggestion in the PR.

This treats falsy but non-nullish inputs, such as Number 0, as valid.

Other than digit zero `0`, I can't think of a falsy JS value that
seems like a valid app name, so they're *probably* misusing the
branding API at this point, but this is for whoever uses `0` as their
`branding.name`. (Note: Please use string `"0"`, or any non-empty
string for that matter, instead!)