deps: Update keytar to 7.9.0 (drops nan --> N-API)
This dependency bump is motivated by seeking to get rid of outdated
nan, which was an indirect dependency via keytar. Dropping nan should
hopefully allow compiling native C/C++ code against newer and newer
versions of NodeJS and Electron -- "hopefully".

This version of keytar drops its dependency on nan, instead using
"N-API", now officially renamed to "Node-API". Hopefully this means
it will not need changes to maintain forward-compatibility with future
versions of NodeJS and Electron -- "hopefully."

So, there may also be benefits for upgrading to newer versions of
Electron -- I haven't tested this, I just want to be able to bootstrap
Pulsar core repo on newer system versions of NodeJS, however
irrelevant system NodeJS version may be to final execution in
Electron.

...

This is the latest, and final upstream release of atom/node-keytar;
The upstream repository is now archived.

If we want to see changes in keytar beyond this, we must search for a
suitable, existing fork, or create one ourselves.