deps: Get whats-my-line as a tarball URL, not a git URL
Also update to a commit of whats-my-line where superstring
is specified as a tarball URL, not a git URL.

Otherwise, no changes.

Just using a simpler format (tarballs), which are quicker
to fetch and quicker to install than git URL dependencies,
as well as being just a lot simpler to deal with under the hood.