deps: Update whats-my-line to latest commit
This is just the commit from *after merging* the whats-my-line PR
that specified superstring as a tarball URL rather than as a git URL.

No code changed, this just points to the merge commit after merging,
instead of the PR branch tip pre-merge.

(For the benefit of folks who spelunk through the commit histories.
This is a slightly tidier landmark in the whats-my-line repo
to point to: a merge commit into `master` branch,
rather than a random-looking pre-merge commit. No change otherwise.)