meta: Delete atomTranspilers from package.json
This prevents the editor for trying to transpile this package
at startup time, which is unnecessary and redundant on this
"pretranspiled" tag of the package. Should save some CPU cycles.

(Also stops an error message from showing that a devDependency was
missing, and that require()ing it failed. This tag may potentially be
requested as a tarball, in which case the devDependencies *aren't
supposed* to be available. Since we genuinely don't need the
devDependencies in production, allowing the package to load without
errors when the devDependencies aren't present is a *good thing*.
Can save us whatever file size the devDependencies take up on disk.)

(cherry picked from commit 32c6b9a98a99a339b1a7508eaee8e173b23959be)