lib: Allow parent scopes in lieu of exact ones
I'd rather folks not put more scopes on their tokens
than they actually need... But the fact this doesn't work is confusing.

We've had to troubleshoot this scenario with users multiple times.
And that's just the ones who reached out about it. Others are likely
affected and unsure how to log in.

This allows supersets of the actually required scopes.
Note to users: Please just set the actual needed scopes...