lib: Pre-transpile the lib/ JS files
This needs to be done at some point before the package is included
in the final editor app bundle, or else the package fails to run.

---

STEPS (how to perform this pre-transpilation):

% npm install --legacy-peer-deps
[ . . . ]
added XXXX packages, and audited X packages in Xm
[ . . . ]

% npm install -g @babel/cli @babel/core
[ . . . ]
changed XX packages in Xs
[ . . . ]

% npm ls -g
[ . . . ]
├── @babel/cli@7.23.4
├── @babel/core@7.23.6
[ . . . ]

% npx babel --version
7.23.4 (@babel/core 7.23.6)

% ELECTRON_VERSION=12 babel lib -d lib --compact auto --source-type module
Successfully compiled 336 files with Babel (XXXXXms).

% git add lib

% git commit [ . . . ]